make server use ts
