+++ package.json
@@ -9,7 +9,8 @@
     "@material-ui/core": "^4.9.14",
     "@types/jest": "^25.2.1",
     "@types/lodash": "^4.14.150",
-    "@types/node": "^13.13.5",
+    "@types/morgan": "^1.9.2",
+    "@types/node": "^14.14.14",
     "@types/querystringify": "^2.0.0",
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.7",
@@ -48,7 +49,7 @@
     "socket.io-client": "^2.3.0",
     "superagent": "^5.0.2",
     "sweetalert": "^2.1.2",
-    "typescript": "^3.8.3",
+    "typescript": "^4.1.3",
     "utility-types": "^3.10.0"
   },
   "type": "module",
@@ -57,15 +58,21 @@
     "@babel/node": "^7.10.5",
     "@babel/preset-env": "^7.11.0",
     "@types/argparse": "^1.0.38",
+    "@types/cors": "^2.8.9",
+    "@types/express": "^4.17.9",
+    "@types/pg": "^7.14.7",
+    "@types/socket.io": "^2.1.12",
+    "@types/uuid": "^8.3.0",
     "argparse": "^1.0.10",
     "husky": "^1.0.0-rc.15",
     "prettier": "^2.0.5",
-    "pretty-quick": "^1.7.0"
+    "pretty-quick": "^1.7.0",
+    "ts-node": "^9.1.1"
   },
   "scripts": {
     "start": "PORT=3020 react-scripts start",
     "devfrontend": "REACT_APP_USE_LOCAL_SERVER=1 PORT=3020 react-scripts start",
-    "devbackend": "PORT=3021 nodemon --exec npx babel-node -- server/server.js",
+    "devbackend": "PORT=3021 nodemon --exec npx babel-node -- server/server.ts",
     "servebackendprod": "while true; do env $(cat .env.prod | xargs) PORT=3021 babel-node server/server.js || true; done",
     "servebackendstaging": "while true; do env $(cat .env.staging | xargs) PORT=4021 babel-node server/server.js || true; done",
     "build": "react-scripts build",
