+++ server/api/puzzle.ts
@@ -1,0 +1,15 @@
+import express from 'express';
+
+import {addPuzzle} from '../model/puzzle';
+
+const router = express.Router();
+router.post('/', async (req, res) => {
+  console.log('got req', req.headers, req.body);
+  addPuzzle(req.body.puzzle, req.body.isPublic);
+  const pid = 123;
+  res.json({
+    pid,
+  });
+});
+
+export default router;
