+++ server/model/game.ts
@@ -1,0 +1,28 @@
+import _ from 'lodash';
+import {pool} from './pool';
+
+export async function getEvents(gid: string) {
+  const startTime = Date.now();
+  const res = await pool.query('SELECT event_payload FROM game_events WHERE gid=$1', [gid]);
+  const events = _.map(res.rows, 'event_payload');
+  const ms = Date.now() - startTime;
+  console.log(`getEvents(${gid}) took ${ms}ms`);
+  return events;
+}
+
+export interface GameEvent {
+  user: any;
+  timestamp: Date;
+  type: any;
+}
+export async function addEvent(gid: string, event: GameEvent) {
+  const startTime = Date.now();
+  await pool.query(
+    `
+      INSERT INTO game_events (gid, uid, ts, event_type, event_payload)
+      VALUES ($1, $2, $3, $4, $5)`,
+    [gid, event.user, new Date(event.timestamp).toISOString(), event.type, event]
+  );
+  const ms = Date.now() - startTime;
+  console.log(`addEvent(${gid}, ${event.type}) took ${ms}ms`);
+}
