+++ server/model/pool.ts
@@ -1,0 +1,9 @@
+const pg = require('pg');
+// ============= Database Operations ============
+
+export const pool = new pg.Pool({
+  host: process.env.PGHOST || 'localhost',
+  user: process.env.PGUSER || process.env.USER,
+  password: process.env.PGPASSWORD,
+  database: process.env.PGDATABASE,
+});
