+++ src/store/api.ts
@@ -71,12 +71,14 @@ export interface PuzzleList {
     numSolves: number;
   };
 }
+[];
 
-export async function fetchPuzzleList(query: {page: number; pageSize: number}) {
-  const url = `${SERVER_URL}/api/puzzles?${qs.stringify(query)}`;
+export async function fetchPuzzleList(query: {page: number; pageSize: number}): Promise<PuzzleList> {
+  const url = `${SERVER_URL}/api/puzzle_list?${qs.stringify(query)}`;
   const resp = await fetch(url);
   const json = await resp.json();
   const puzzleList: PuzzleList = json;
+
   return puzzleList;
 }
