+++ tsconfig.json
@@ -1,22 +1,8 @@
 {
   "compilerOptions": {
-    "outDir": "./dist/",
-    "sourceMap": true,
+    "noImplicitAny": true,
     "strictNullChecks": true,
-    "module": "esnext",
-    "jsx": "react",
-    "target": "es5",
-    "allowJs": true,
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true
+    "esModuleInterop": true
   },
-  "include": ["./src/"]
+  "include": ["src", "server"]
 }
